Proxy Model
Added the 'ProxyModel.cs' class which handles a game object which either
displays a primitive sphere, or a loaded mesh.

The Resource manager now creates or destroys proxy model objects when
changes happen in the streaming assets directory.

Added a wood finish to the table and a metallic look to the legs.